+++ Sources/CodableWrappers/CodableMacros/CodableMacros.swift
@@ -5,65 +5,65 @@
 import CodableWrapperMacros
 /// Add to a struct to enable custom CodingKeys
 @attached(member, names: named(CodingKeys))
-public macro CCodable() = #externalMacro(module: "CodableWrapperMacros", type: "CCodable")
+public macro CustomCodable() = #externalMacro(module: "CodableWrapperMacros", type: "CustomCodable")
 
-/// Use a custom String value for a Property's CodingKey
+/// Set a property's CodingKey with a custom value
 @attached(peer)
-public macro CodingKey(_ name: StringLiteralType) = #externalMacro(module: "CodableWrapperMacros", type: "CodingKey")
+public macro CustomCodingKey(_ name: StringLiteralType) = #externalMacro(module: "CodableWrapperMacros", type: "CustomCodingKey")
 
-/// Prefix the CodingKey with a custom value
+/// Add a Prefix a property's CodingKey or Type's CodingKeys with a custom value
 @attached(peer)
 public macro CodingKeyPrefix(_ name: StringLiteralType) = #externalMacro(module: "CodableWrapperMacros", type: "CodingKeyPrefix")
 
-/// Add a custom Suffix to the CodingKey
+/// Add a Suffix a property's CodingKey or Type's CodingKeys with a custom value
 @attached(peer)
 public macro CodingKeySuffix(_ name: StringLiteralType) = #externalMacro(module: "CodableWrapperMacros", type: "CodingKeySuffix")
 
-/// CodingKey value will be `camelCase`
+/// Make the CodingKey for a property or Type `camelCase`
 @attached(peer)
 public macro CamelCase() = #externalMacro(module: "CodableWrapperMacros", type: "CamelCase")
 
-/// CodingKey value will be` flatcase`
+/// Make the CodingKey for a property or Type` flatcase`
 @attached(peer)
 public macro FlatCase() = #externalMacro(module: "CodableWrapperMacros", type: "FlatCase")
 
-/// CodingKey value will be `PascalCase`
+/// Make the CodingKey for a property or Type `PascalCase`
 @attached(peer)
 public macro PascalCase() = #externalMacro(module: "CodableWrapperMacros", type: "PascalCase")
 
-/// CodingKey value will be `UPPERCASE`
+/// Make the CodingKey for a property or Type `UPPERCASE`
 @attached(peer)
 public macro UpperCase() = #externalMacro(module: "CodableWrapperMacros", type: "UpperCase")
 
-/// CodingKey value will be `snake_case`
+/// Make the CodingKey for a property or Type `snake_case`
 @attached(peer)
 public macro SnakeCase() = #externalMacro(module: "CodableWrapperMacros", type: "SnakeCase")
 
-/// CodingKey value will be `camel_Snake_Case`
+/// Make the CodingKey for a property or Type `camel_Snake_Case`
 @attached(peer)
 public macro CamelSnakeCase() = #externalMacro(module: "CodableWrapperMacros", type: "CamelSnakeCase")
 
-/// CodingKey value will be `Pascal_Snake_Case`
+/// Make the CodingKey for a property or Type `Pascal_Snake_Case`
 @attached(peer)
 public macro PascalSnakeCase() = #externalMacro(module: "CodableWrapperMacros", type: "PascalSnakeCase")
 
-/// CodingKey value will be `SCREAMING_SNAKE_CASE`
+/// Make the CodingKey for a property or Type `SCREAMING_SNAKE_CASE`
 @attached(peer)
 public macro ScreamingSnakeCase() = #externalMacro(module: "CodableWrapperMacros", type: "ScreamingSnakeCase")
 
-/// CodingKey value will be `kebab-case`
+/// Make the CodingKey for a property or Type `kebab-case`
 @attached(peer)
 public macro KebabCase() = #externalMacro(module: "CodableWrapperMacros", type: "KebabCase")
 
-/// CodingKey value will be `camel-Kebab-Case`
+/// Make the CodingKey for a property or Type `camel-Kebab-Case`
 @attached(peer)
 public macro CamelKebabCase() = #externalMacro(module: "CodableWrapperMacros", type: "CamelKebabCase")
 
-/// CodingKey value will be `Pascal-Kebab-Case`
+/// Make the CodingKey for a property or Type `Pascal-Kebab-Case`
 @attached(peer)
 public macro PascalKebabCase() = #externalMacro(module: "CodableWrapperMacros", type: "PascalKebabCase")
 
-/// CodingKey value will be `SCREAMING-KEBAB-CASE`
+/// Make the CodingKey for a property or Type `SCREAMING-KEBAB-CASE`
 @attached(peer)
 public macro ScreamingKebabCase() = #externalMacro(module: "CodableWrapperMacros", type: "ScreamingKebabCase")