+++ Tests/IntegrationTests/IntegrationTests.swift
@@ -27,17 +27,9 @@ final class IntegrationTests: XCTestCase {
         // Mark your test throws to produce an unexpected failure when your test encounters an uncaught error.
         // Mark your test async to allow awaiting for asynchronous code to complete. Check the results with assertions afterwards.
     }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
 }
 //
-//@CCodable @SnakeCase
+//@CustomCodable @SnakeCase
 //struct User: Codable {
 //    let firstName: String
 //    let lastName: String